requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pydantic-settings==2.3.4

# PDF 직접 추출
pdfplumber==0.11.4

# PDF 페이지 → 이미지 변환 (시스템에 poppler 필요)
pdf2image==1.17.0

# 이미지 전처리 및 품질 평가
opencv-python-headless==4.11.0.86

# OCR 엔진
pytesseract==0.3.13
# 시스템에 Tesseract 5.x + kor+eng 언어팩이 설치되어 있어야 함

# 이미지 파일 처리
easyocr==1.7.1
Pillow==10.4.0
numpy==1.26.4

# OCR 신뢰도 데이터 처리
pandas==2.2.2

pytest==8.3.2
httpx==0.27.2
